--- app/(main)/order/form/_components/payment_info_section.module.css
@@ -1,3 +1,0 @@
-.payment_info_section_checkout__button {
-  margin: 100px auto 0;
-}
