+++ app/(main)/order/form/_components/payment_section.module.css
@@ -1,0 +1,3 @@
+.payment_section_checkout__button {
+  margin: 100px auto 0;
+}
